+++ src/test/java/com/example/green/domain/point/controller/PointTransactionControllerTest.java
@@ -15,10 +15,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 
-import com.example.green.domain.point.controller.dto.MemberPointSummary;
-import com.example.green.domain.point.controller.dto.MyPointTransaction;
-import com.example.green.domain.point.controller.query.PointTransactionQueryRepository;
 import com.example.green.domain.point.entity.vo.TransactionType;
+import com.example.green.domain.point.repository.PointTransactionQueryRepository;
+import com.example.green.domain.point.repository.dto.MemberPointSummary;
+import com.example.green.domain.point.repository.dto.MyPointTransaction;
 import com.example.green.global.api.ApiTemplate;
 import com.example.green.global.api.page.CursorTemplate;
 import com.example.green.template.base.BaseControllerUnitTest;
